IniFile: Use more templates.

+++ Source/Core/Core/HW/WiimoteReal/WiimoteReal.cpp
@@ -622,12 +622,12 @@ void LoadSettings()
     secname += (char)('1' + i);
     IniFile::Section& sec = *inifile.GetOrCreateSection(secname);
 
-    sec.Get("Source", &g_wiimote_sources[i], i ? WIIMOTE_SRC_NONE : WIIMOTE_SRC_EMU);
+    sec.Get("Source", &g_wiimote_sources[i], u32(i ? WIIMOTE_SRC_NONE : WIIMOTE_SRC_EMU));
   }
 
   std::string secname("BalanceBoard");
   IniFile::Section& sec = *inifile.GetOrCreateSection(secname);
-  sec.Get("Source", &g_wiimote_sources[WIIMOTE_BALANCE_BOARD], WIIMOTE_SRC_NONE);
+  sec.Get("Source", &g_wiimote_sources[WIIMOTE_BALANCE_BOARD], (u32)WIIMOTE_SRC_NONE);
 }
 
 // config dialog calls this when some settings change
